requirements.txt
boto3==1.5.2
fabric3==1.13.1.post1
